+++ slices/reverse.go
@@ -1,7 +1,6 @@
 package slices
 
 import (
-	"errors"
 	"reflect"
 
 	"github.com/golodash/godash/internal"
@@ -10,14 +9,11 @@ import (
 func Reverse(slice interface{}) (interface{}, error) {
 	err := internal.SliceCheck(slice)
 	if err != nil {
-		return nil, errors.New("the entry should be slice")
+		return nil, err
 	}
-	n := reflect.ValueOf(slice).Len()
-	if n == 0 {
-		return nil, nil
-	}
+	length := reflect.ValueOf(slice).Len()
 	swap := reflect.Swapper(slice)
-	for i, j := 0, n-1; i < j; i, j = i+1, j-1 {
+	for i, j := 0, length-1; i < j; i, j = i+1, j-1 {
 		swap(i, j)
 	}
 	return slice, nil